map refactoring - structure done, functionality needs update
